ingestor/cloud_functions/gemini_extractor/requirements.txt
google-cloud-firestore
google-generativeai
pyyaml 